refactor: Break out database definitions from models into shared helper.
