+++ tools/serverpod_cli/test/database/extentions/migration_pgsql_code_generation_test.dart
@@ -1,15 +1,10 @@
 import 'package:serverpod_cli/analyzer.dart';
-import 'package:serverpod_cli/src/analyzer/models/stateful_analyzer.dart';
-import 'package:serverpod_cli/src/database/create_definition.dart';
-import 'package:serverpod_cli/src/generator/code_generation_collector.dart';
-import 'package:serverpod_cli/src/util/model_helper.dart';
-import 'package:serverpod_service_client/serverpod_service_client.dart';
 import 'package:test/test.dart';
 
-import '../../generator/dart/client_code_generator/class_constructors_test.dart';
 import '../../test_util/builders/database/database_definition_builder.dart';
 import '../../test_util/builders/database/table_definition_builder.dart';
 import '../../test_util/builders/model_source_builder.dart';
+import '../../test_util/database_definition_helpers.dart';
 
 void main() {
   group(
@@ -82,8 +77,7 @@ fields:
       ).build(),
     ];
 
-    var (:sourceDefinition, :targetDefinition) =
-        _createDatabaseDefinitionFromModels(
+    var (:sourceDefinition, :targetDefinition) = databaseDefinitionsFromModels(
       sourceModels,
       targetModels,
     );
@@ -104,29 +98,4 @@ fields:
 
     expect(createNewModelTable, lessThan(addForeignKeyToExistingTable));
   });
-}
-
-({DatabaseDefinition sourceDefinition, DatabaseDefinition targetDefinition})
-    _createDatabaseDefinitionFromModels(
-  List<ModelSource> sourceModels,
-  List<ModelSource> targetModels,
-) {
-  DatabaseDefinition createDefinition(models) {
-    var collector = CodeGenerationCollector();
-    var analyzer =
-        StatefulAnalyzer(config, models, onErrorsCollector(collector));
-    var definitions = analyzer.validateAll();
-    expect(collector.errors, isEmpty);
-
-    return createDatabaseDefinitionFromModels(
-      definitions,
-      'example',
-      [],
-    );
-  }
-
-  return (
-    sourceDefinition: createDefinition(sourceModels),
-    targetDefinition: createDefinition(targetModels)
-  );
 }
